AKU-812: Commas not valid in NumberSpinner - Updates following code review
